fix(online): redis->NonBlockingHashMap

+++ src/main/java/cn/edu/zut/mfs/service/UserService.java
@@ -56,5 +56,5 @@ public interface UserService {
 
     List<User> onlineList();
 
-    int onlineUsers();
+    Long onlineUsers();
 }
